requirements.txt
fastapi[all]==0.104.1
SQLAlchemy==2.0.23
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
